requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
numpy==2.2.0
huggingface-hub==0.26.5

# Database (for vocabulary feature)
sqlalchemy>=2.0
psycopg[binary]>=3.0
pgvector
alembic>=1.13.0

# For music generation lab (pt_part2_music_generation.py)
torch
torchaudio
mitdeeplearning
tqdm
scipy
opencv-python

# For TensorFlow music generation lab
tensorflow
